requirements.txt
streamlit
audio-recorder-streamlit
# Utilitites 
numpy
pandas

# FACE RECOGNITION
scikit-learn
dlib-bin
opencv-python
git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0


supabase
bcrypt



segno
pillow



librosa
resemblyzer